requirements.txt
streamlit
pdfplumber
pikepdf
pytesseract
Pillow
python-dateutil
